feature: unified send flow for custom and default tokens
